refactor: move the runtime scope primitives above pysepal.solara
The i18n package needs per-scope state. It must not import pysepal.solara,
because that module also imports the widgets. The old modules re-export the
primitives, so every existing import path still works.
