+++ tests/test_solara/test_notifications/test_globals.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 import pytest
 
-from pysepal.solara import scope_registry
+from pysepal import _scope_registry as scope_registry
 from pysepal.solara.notifications.bus import _refcounts, _registry
 from pysepal.solara.notifications.state import TaskStatus, ToastType
